requirements.txt
besser-bot-framework==1.0.0
langchain==0.0.344
langchain-core==0.1.12
langchain-experimental==0.0.42
langchain-openai==0.0.2.post1
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.1
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.8.0